:bug: disable comment when request in progress n handle setState error
