+++ lib/pages/home/pages/post_comments/widgets/post_comment/post_comment.dart
@@ -2,15 +2,18 @@ import 'package:Openbook/models/community.dart';
 import 'package:Openbook/models/post.dart';
 import 'package:Openbook/models/post_comment.dart';
 import 'package:Openbook/models/user.dart';
-import 'package:Openbook/pages/home/pages/post_comments/widgets/post_comment/packages/post_comment_text.dart';
-import 'package:Openbook/pages/home/pages/post_comments/widgets/post_comment/post_comment_actions.dart';
+import 'package:Openbook/pages/home/pages/post_comments/widgets/post_comment/widgets/post_comment_text.dart';
 import 'package:Openbook/provider.dart';
-import 'package:Openbook/services/bottom_sheet.dart';
+import 'package:Openbook/services/modal_service.dart';
 import 'package:Openbook/services/navigation_service.dart';
+import 'package:Openbook/services/toast.dart';
+import 'package:Openbook/services/user.dart';
 import 'package:Openbook/widgets/avatars/avatar.dart';
 import 'package:Openbook/widgets/icon.dart';
 import 'package:Openbook/widgets/theming/secondary_text.dart';
+import 'package:async/async.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_slidable/flutter_slidable.dart';
 
 class OBPostComment extends StatefulWidget {
   final PostComment postComment;
@@ -32,31 +35,48 @@ class OBPostComment extends StatefulWidget {
 
 class OBPostCommentState extends State<OBPostComment> {
   NavigationService _navigationService;
-  BottomSheetService _bottomSheetService;
+  UserService _userService;
+  ToastService _toastService;
+  ModalService _modalService;
+  bool _requestInProgress;
+
+  CancelableOperation _requestOperation;
 
   @override
   void initState() {
     super.initState();
+    _requestInProgress = false;
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    if (_requestOperation != null) _requestOperation.cancel();
   }
 
   @override
   Widget build(BuildContext context) {
     var provider = OpenbookProvider.of(context);
     _navigationService = provider.navigationService;
-    _bottomSheetService = provider.bottomSheetService;
+    _userService = provider.userService;
+    _toastService = provider.toastService;
+    _modalService = provider.modalService;
     Widget postTile = _buildPostCommentTile(widget.postComment);
 
-    return OBPostCommentActions(
-      post: widget.post,
-      postComment: widget.postComment,
-      onPostCommentDeletedCallback: widget.onPostCommentDeletedCallback,
+    Widget postComment = _buildPostCommentActions(
       child: postTile,
     );
-  }
 
-  @override
-  void dispose() {
-    super.dispose();
+    if (_requestInProgress) {
+      postComment = IgnorePointer(
+        child: Opacity(
+          opacity: 0.5,
+          child: postComment,
+        ),
+      );
+    }
+
+    return postComment;
   }
 
   Widget _buildPostCommentTile(PostComment postComment) {
@@ -108,6 +128,104 @@ class OBPostCommentState extends State<OBPostComment> {
             ),
           );
         });
+  }
+
+  Widget _buildPostCommentActions({@required Widget child}) {
+    List<Widget> _editCommentActions = [];
+
+    User loggedInUser = _userService.getLoggedInUser();
+    bool loggedInUserIsCommunityAdministrator = false;
+    bool loggedInUserIsCommunityModerator = false;
+
+    Post post = widget.post;
+    User postCommenter = widget.postComment.commenter;
+
+    if (post.hasCommunity()) {
+      Community postCommunity = post.community;
+
+      loggedInUserIsCommunityAdministrator =
+          postCommunity.isAdministrator(loggedInUser);
+
+      loggedInUserIsCommunityModerator =
+          postCommunity.isModerator(loggedInUser);
+    }
+
+    if (postCommenter.id == loggedInUser.id) {
+      _editCommentActions.add(
+        new IconSlideAction(
+          caption: 'Edit',
+          color: Colors.blueGrey,
+          icon: Icons.edit,
+          onTap: _editPostComment,
+        ),
+      );
+    }
+
+    if (widget.postComment.getCommenterId() == loggedInUser.id ||
+        loggedInUserIsCommunityAdministrator ||
+        loggedInUserIsCommunityModerator ||
+        post.creator.id == loggedInUser.id) {
+      _editCommentActions.add(
+        new IconSlideAction(
+          caption: 'Delete',
+          color: Colors.red,
+          icon: Icons.delete,
+          onTap: _deletePostComment,
+        ),
+      );
+    }
+
+    return Slidable(
+      delegate: new SlidableDrawerDelegate(),
+      actionExtentRatio: 0.2,
+      child: child,
+      secondaryActions: _editCommentActions,
+    );
+  }
+
+  void _editPostComment() async {
+    await _modalService.openExpandedCommenter(
+        context: context, post: widget.post, postComment: widget.postComment);
+  }
+
+  void _deletePostComment() async {
+    if (_requestInProgress) return;
+    _setRequestInProgress(true);
+    try {
+      _requestOperation = CancelableOperation.fromFuture(
+          _userService.deletePostComment(
+              postComment: widget.postComment, post: widget.post));
+
+      await _requestOperation.value;
+      widget.post.decreaseCommentsCount();
+      _toastService.success(message: 'Comment deleted', context: context);
+      if (widget.onPostCommentDeletedCallback != null) {
+        widget.onPostCommentDeletedCallback();
+      }
+    } catch (error) {
+      _onError(error);
+    } finally {
+      _setRequestInProgress(false);
+    }
+  }
+
+  void _setRequestInProgress(bool requestInProgress) {
+    setState(() {
+      _requestInProgress = requestInProgress;
+    });
+  }
+
+  void _onError(error) async {
+    if (error is HttpieConnectionRefusedError) {
+      _toastService.error(
+          message: error.toHumanReadableMessage(), context: context);
+    } else if (error is HttpieRequestError) {
+      String errorMessage = await error.toHumanReadableMessage();
+      _toastService.error(message: errorMessage, context: context);
+    } else {
+      _toastService.error(message: 'Unknown error', context: context);
+      throw error;
+    }
   }
 
   Widget _getCommunityBadge(PostComment postComment) {
